changelog previous workhome changed change transferred network methods from JsonTMDBLoad.kt to MoviesNetwork.kt and MovieApiInterface.kt delete JsonTMDBLoad.kt add constructor for MoviesNetwork, MoviesRepository, update MoviesViewModelFactory, DetailsMovieViewModelFactory
