--- app/src/main/java/ru/gorinih/androidacademy/data/network/MovieNetwork.kt
@@ -1,11 +1,0 @@
-package ru.gorinih.androidacademy.data.network
-
-import ru.gorinih.androidacademy.data.loadMovieById
-import ru.gorinih.androidacademy.data.model.Movies
-import ru.gorinih.androidacademy.data.loadMovies
-
-class MovieNetwork {
-    suspend fun getMoviesNet(numberPage: Int): List<Movies.Movie> = loadMovies(numberPage)
-
-    suspend fun getMovieDetailsNet(id: Int): Movies.Movie? = loadMovieById(id)
-}
