+++ app/src/main/java/ru/gorinih/androidacademy/viewmodel/MoviesViewModel.kt
@@ -7,12 +7,18 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
+import kotlinx.serialization.ExperimentalSerializationApi
 import ru.gorinih.androidacademy.data.model.Movies
 import ru.gorinih.androidacademy.data.repository.MoviesRepository
 import kotlin.coroutines.CoroutineContext
 
+@ExperimentalSerializationApi
 class MoviesViewModel(private val moviesRepository: MoviesRepository) : ViewModel() {
 
     private var page: Int = 0
@@ -29,22 +35,32 @@ class MoviesViewModel(private val moviesRepository: MoviesRepository) : ViewMode
     val movieList: LiveData<List<Movies>>
         get() = _movieList
 
+    lateinit var flowMovie: Flow<List<Movies>>
+
     init {
         nextMovies()
     }
 
     private fun getMoviesList(numberPage: Int) {
         viewModelScope.launch(exceptionMovie) {
-            val listMovies = mutableListOf<Movies>()
             val oldMovies = movieList.value ?: listOf(Movies.Header)
             val listIdMovies = oldMovies.filterIsInstance(Movies.Movie::class.java).map { it.id }
-            listMovies.addAll(oldMovies)
+/*
+           flowMovie = moviesRepository.getMovies(numberPage, listIdMovies)
+           flowMovie.collect {
+               _movieList.value = it
+           }
+*/
             val newMovies = moviesRepository.getMovies(numberPage, listIdMovies)
             if (newMovies.isNotEmpty()) {
-                refreshCache(newMovies)
+                val listMovies = mutableListOf<Movies>()
+                listMovies.addAll(oldMovies)
+                refreshCache(newMovies) //- ее в репозитории аналог пишем
+                listMovies.addAll(newMovies)
+                _movieList.value = listMovies
+            } else {
+                page--
             }
-            listMovies.addAll(newMovies)
-            _movieList.value = listMovies
         }
     }
